// src/config/constants/types.ts
export interface Address {
  97?: string
  56: string
}
// ...
export interface BountyConfig {
  id: number,
  contractAddress: Address,
  name: string,
  isPremium?: boolean,
  title?: string
  description?: string
  image: string
}